Configure Spring Batch with execution listener

+++ server/team-space-data-import/src/main/java/org/teamspace/data_import/job_config/tasks/DataWriterTasklet.java
@@ -8,10 +8,12 @@ import org.springframework.batch.core.step.tasklet.Tasklet;
 import org.springframework.batch.repeat.RepeatStatus;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.teamspace.auth.domain.User;
+import org.teamspace.data_import.domain.DataImportRequest;
 import org.teamspace.users.service.UsersService;
 
 import java.util.List;
 
+import static org.teamspace.data_import.constants.DataImportConstants.CUSTOM_PARAMETERS_JOB_KEY;
 import static org.teamspace.data_import.constants.DataImportConstants.USERS_DATA_JOB_KEY;
 
 /**
@@ -28,7 +30,11 @@ public class DataWriterTasklet implements Tasklet {
         log.info("Started import users to database");
         JobExecution jobExecution = chunkContext.getStepContext().getStepExecution().getJobExecution();
         List<User> users = (List<User>)jobExecution.getExecutionContext().get(USERS_DATA_JOB_KEY);
-        usersService.deleteNonAdminUsers();
+        DataImportRequest dataImportRequest = (DataImportRequest)jobExecution
+                .getExecutionContext().get(CUSTOM_PARAMETERS_JOB_KEY);
+        if(dataImportRequest.isShouldPerformCleanup()) {
+            usersService.deleteNonAdminUsers();
+        }
         usersService.importUsers(users);
         log.info("Completed import users to database");
         return RepeatStatus.FINISHED;
